+++ test/re_frame_datatable/subs_test.clj
@@ -3,7 +3,8 @@
              [core :as dt]
              [subs :as sut]
              [paths :as p]
-             [db :as db]]
+             [db :as db]
+             [utils :as u]]
             [midje.sweet :refer :all]
             [re-frame
              [core :as rf]
@@ -92,6 +93,7 @@
            (fact "filters data case sensitive"
              (reset! app-db (-> {}
                               (assoc ::items [{:name "Item 1"} {:name "item 2"}])
+                              (db/set-options id {::dt/filtering {::dt/enabled? true}})
                               (db/set-filtering id {:search-phrase "item"}))) => map?
              (count (:visible-items @d)) => 1
              (:visible-items @d) => [[1 {:name "item 2"}]])
@@ -99,7 +101,8 @@
            (fact "filters data case insensitive"
              (reset! app-db (-> {}
                               (assoc ::items [{:name "Item 1"} {:name "item 2"}])
-                              (db/set-options id {::dt/filtering {::dt/case-insensitive-filtering true}})
+                              (db/set-options id {::dt/filtering {::dt/enabled? true
+                                                                  ::dt/filtering-fn u/case-insensitive-filtering-fn}})
                               (db/set-filtering id {:search-phrase "item"}))) => map?
              (count (:visible-items @d)) => 2
              (:visible-items @d) => [[0 {:name "Item 1"}] [1 {:name "item 2"}]]))))